AppAdmin interface expansion commit 5: can now add new homes. Final commit in this saga.
--
Add home management features in AdminController and UI

Added logic in `AdminController.cs` to handle the removal of `JoinCommunityRequest` objects when a home is removed from a community. Introduced a new API endpoint `AdminAddHome` with necessary validation checks. Updated `ApiEndPoints.cs` to include this new endpoint.

Enhanced `AppAdmin.razor` UI to include a form for adding a home to a community. Added a new `HomeDTO` object `HomeDTOToAdd` in `AppAdmin.razor.cs` to hold the data for the home being added. Implemented the `AddHome` method to handle the addition logic and UI updates. Refactored the `EditHome` method to reset the `HomeDTOToEdit` object only upon successful operation.