Removed Include Authors when searching for Post since it is of type interface IAvatar and not a concrete type which was causing issues.
--
Remove eager loading of Author in Post query

The `Include(p => p.Author)` statement was removed from the Entity Framework query in `CommunityController.cs` that retrieves a `Post` object. This change means that the `Author` property of the `Post` will no longer be eagerly loaded when fetching the post from the database.